init
This branch will reconstruct the parser ind Model.java.
This is done parse multible times and then easing the load of objects created at once.
Furthermore this will organize the parsing a lot more and futureproof the model.
This was originally made because of problems with coloring the structures, because of bad parse handling.
- Rasmus A. Chr.