Changed to look more like before Borderpane change
Added a pane to be the center of the borderpane. This pane (called backgroundPane) then contains the mapPane and the searchBar. This keeps the benefits of the borderpane, whilst looking more like the desired look of the design.

Known issue: The mapPane only resizes when dragging the mouse inside that pane, and not the backgroundPane