Added center address panning and bigger range for rangesearch
panToAddress now pans over to the address and also puts the address in the middle of canvas. It also draws a circle on the address, so that the user more easily can see where the address actually is.

Has also made the range for rangesearch a bit bigger, so that we experience less popins than before.